LayoutTests/imported/w3c/web-platform-tests/content-security-policy/navigate-to/href-location-blocked.sub-expected.txt


FAIL Test that the child iframe navigation is not allowed assert_equals: expected "fail" but got "success"
FAIL Violation report status OK. undefined is not an object (evaluating 'data[0]["body"]')
